[Feat] String Extension 으로 분리 (#7)
