+++ assignment1/assignment1/Presentations/WeatherList/WetherListTableViewCell.swift
@@ -89,16 +89,6 @@ class WetherListTableViewCell: UITableViewCell {
             $0.leading.equalToSuperview().inset(272)
         }
     }
-    //    현재시간을 반환하기 위한 함수입니다.
-    func getCurrentTime() -> String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "HH:mm"
-        
-        let currentTime = Date()
-        let formattedTime = dateFormatter.string(from: currentTime)
-        
-        return formattedTime
-    }
     
     func bindData(data: WeatherListViewData) {
         self.myLocationLabel.text = data.location
